Add(config): mock cors config
