+++ jacoco-exclude-class
@@ -1,3 +1,4 @@
 exclude com/newbarams/ajaja/AjajaApplication
 exclude com/newbarams/ajaja/global/config/*
 exclude com/newbarams/ajaja/global/common/*
+exclude com/newbarams/ajaja/global/mock/*
